renamed data_handers.factory to data_handlers.base, after removal of explicit factory class. Clarified which class were inherited and composed in doc strings.
