+++ sup3r/preprocessing/data_handlers/__init__.py
@@ -1,10 +1,10 @@
 """Composite objects built from loaders, rasterizers, and derivers."""
 
-from .exo import ExoData, ExoDataHandler, SingleExoDataStep
-from .factory import (
+from .base import (
     DailyDataHandler,
     DataHandler,
     DataHandlerH5SolarCC,
     DataHandlerH5WindCC,
 )
+from .exo import ExoData, ExoDataHandler, SingleExoDataStep
 from .nc_cc import DataHandlerNCforCC, DataHandlerNCforCCwithPowerLaw
